+++ cfgov/v1/models/atoms.py
@@ -2,5 +2,5 @@ from wagtail.wagtailcore import blocks
 
 
 class Hyperlink(blocks.StructBlock):
-    label = blocks.CharBlock(max_length=22)
+    label = blocks.CharBlock(max_length=50)
     href = blocks.CharBlock(default='/')